annotate all db entities

+++ backend/src/main/java/de/hanselmann/shoppinglist/restapi/dto/InfoDto.java
@@ -10,19 +10,19 @@ public class InfoDto {
     }
 
     public static class InfoMessageDto {
-        private final int messageNumber;
+        private final long messageNumber;
         private final SeverityDto severity;
         private final String message;
         private final String dismissLabel;
 
-        public InfoMessageDto(int messageNumber, SeverityDto severity, String message, String dismissLabel) {
+        public InfoMessageDto(long messageNumber, SeverityDto severity, String message, String dismissLabel) {
             this.messageNumber = messageNumber;
             this.severity = severity;
             this.message = message;
             this.dismissLabel = dismissLabel;
         }
 
-        public int getMessageNumber() {
+        public long getMessageNumber() {
             return messageNumber;
         }
